Make table change an ALTER and set next sequence to start from 50020

+++ src/main/resources/db/migration/V1.0.1__init_tables.sql
@@ -2,7 +2,7 @@
 CREATE SCHEMA IF NOT EXISTS ia_case_api;
 --
 CREATE TABLE IF NOT EXISTS ia_case_api.appeal_reference_numbers (
-  case_id BIGINT NOT NULL,
+  case_id INT NOT NULL,
   type varchar(30),
   year INT NOT NULL,
   sequence INT,
